BaiTapLon_phat_nguoi/requirements.txt
selenium
Pillow
pytesseract
schedule